% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cols_length.R
\name{cols_length}
\alias{cols_length}
\title{Used to get the max length of each variable in the dataset}
\usage{
cols_length(df, out_name = "col_lengths")
}
\arguments{
\item{df}{Name of the dataset}

\item{out_name}{Name of the output dataset in which max lengths should be stored (optional)}
}
\value{
returns the max length of each variable in a seperate dataframe
}
\description{
Checks the length of each column name in \code{df}. Arguments will be unquoted names,
}
\examples{
pdata <- data.frame(name = c("alice", "barbara", "carol", "darwin"), id = c(1014, 1000, 1012, 1100))
cols_length(pdata)
}
