% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createLmFormula.R
\name{createLmFormula}
\alias{createLmFormula}
\title{createLmFormula}
\usage{
createLmFormula(response, explicatives)
}
\arguments{
\item{response}{Responsable variable in string}

\item{explicatives}{Explicative variables in string array}
}
\description{
createLmFormula
}
