% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{drawAnovaPlot}
\alias{drawAnovaPlot}
\title{drawQualiAndQuanti}
\usage{
drawAnovaPlot(data, paramY, paramX)
}
\arguments{
\item{data}{Dataframe with datas}

\item{paramY}{String with Y parameter name}

\item{paramX}{String with Xs parameter name}
}
\description{
drawQualiAndQuanti
}
