% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{drawQuantiVsQualiQuanti}
\alias{drawQuantiVsQualiQuanti}
\title{drawQuantiVsQualiQuanti}
\usage{
drawQuantiVsQualiQuanti(data, paramY, Xquanti, Xquali)
}
\arguments{
\item{data}{Dataframe with datas}

\item{paramY}{Name of response quantitative parameter}

\item{Xquanti}{Name of quantitative parameter}

\item{Xquali}{Name of qualitative parameter}
}
\description{
drawQuantiVsQualiQuanti
}
