% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{drawRegressionPlot}
\alias{drawRegressionPlot}
\title{drawRegressionPlot}
\usage{
drawRegressionPlot(paramY, paramX, data)
}
\arguments{
\item{paramY}{String with Y parameter name}

\item{paramX}{String with X parameter name}

\item{data}{Dataframe with datas}
}
\description{
drawRegressionPlot
}
