% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_dup_out.R
\name{find_dup_out}
\alias{find_dup_out}
\title{Check and handle duplicate rows in a dataset}
\usage{
find_dup_out(df, dupout = "duplicates", cleandup = "nodup_df")
}
\arguments{
\item{df}{A dataframe or tibble.}

\item{dupout}{A character string. The name of the dataset in which duplicates will be stored.
The default is \code{"duplicates"}.}

\item{cleandup}{A character string. The name of the dataset in which only non-duplicate rows will be stored.
The default is \code{"nodup_df"}.}
}
\value{
Creates two datasets in the global environment:
\itemize{
\item \code{dupout}: containing duplicate rows.
\item \code{cleandup}: containing the dataset without duplicates.
Also returns the cleaned dataset invisibly.
}
}
\description{
This function checks if your dataset has any duplicate rows.
If duplicates exist, it can remove them and store them in a separate dataset.
}
\note{
Messages will be shown such as:
\itemize{
\item "Your dataset 'dataframe' has 1 duplicate row."
\item "Removing them and storing the duplicates in a dataset called 'duplicates'."
}
}
\examples{
dataframe <- data.frame(
  apha = c("a", "a", "b", "c"),
  num = c(1, 1, 2, 3),
  symbols = c("$", "$", "8", "#")
)
find_dup_out(dataframe)
find_dup_out(dataframe, "dup", "nodup")

}
