% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getModelParametersFromModelName.R
\name{getModelParametersFromModelName}
\alias{getModelParametersFromModelName}
\title{getModelVariablesFromName}
\usage{
getModelParametersFromModelName(modelName)
}
\arguments{
\item{modelName}{Modele name to extract informations from}
}
\description{
getModelVariablesFromName
}
