% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelIsSignificant.R
\name{modelIsSignificant}
\alias{modelIsSignificant}
\title{modelIsSignificant}
\usage{
modelIsSignificant(model)
}
\arguments{
\item{model}{Model linear to test significativity}
}
\value{
TRUE if model is significant, FALSE else
}
\description{
modelIsSignificant
}
